[mvvm] extract a factories for test
